requirements.txt
streamlit>=1.30.0
google-search-results>=2.4.2
google-genai>=0.3.2
agno>=0.1.5
python-dotenv>=1.0.1
yagmail>=0.15.293

